modifying exisiting functions, seperate script.js to several modules
