+++ data.js
@@ -804,3 +804,6 @@ const agoraStatesDiscussions = [
     bodyHTML: DOMPurify.sanitize(discussion.bodyHTML)
   }
 })
+
+
+export default agoraStatesDiscussions;
