+++ index.html
@@ -74,7 +74,7 @@
     </div>  
   </main>
 </body>
-<script src="data.js"></script>
-<script src="script.js"></script>
+<script type= "module" src="data.js"></script>
+<script type="module" src="App.js"></script>
 
 </html>
